pyproject.toml
[tool.stela]
use_environment_layers = true
environment_variable_name = "ENV"
default_environment = "local"
show_logs = false

[environment]
project.allow_all_origins = true
project.name = "back-edesoft"
project.secret_key = "foo"
project.debug = false
project.log_level = "INFO"
project.service_type = "api"
project.show_django_pages = false
project.auth_api_url = "http://auth_dev"
project.allowed_hosts = ["*"]
project.reload_app = false
project.database_schemas = "public"
project.show_debug_toolbar = false

[environment.local]
project.allow_all_origins = true
project.debug = true
project.log_level = "DEBUG"
project.reload_app = true
project.show_debug_toolbar = true
project.show_django_pages = true
project.testing = true

[tool.black]
target-version = ['py38']
include = '\.pyi?$'
line-length = 120
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | \venv
    | \.aws-sam
    | _build
    | buck-out
    | build
    | dist
    | migrations
    | node_modules
    | dependencies
  )/
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[tool.semantic_release]
version_variable = [
    "core/__init__.py:__version__",
    "pyproject.toml:version"
]
branch = "main"
upload_to_pypi = false
upload_to_release = true
build_command = "python -m pip install -U poetry && poetry build"

[tool.poetry]
name = "back-edesoft"
version = "1.0.0"
description = ""
authors = ["Marco Aurélio <marcodkts@wolfgear.tech>"]
readme = "README.md"
packages = [{include = "back-edesoft"}]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.13.1"
Django = "^4.1.1"
loguru = "^0.6.0"
requests = "^2.28.1"
arrow = "^1.2.3"
stela = "^4.0.2"
stringcase = "^1.2.0"
asbool = "^0.2"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.3"
drf-spectacular = "^0.24.0"
django-cors-headers = "^3.13.0"
PyJWT = "^2.5.0"
celery = "^5.2.7"
django-celery-results = "^2.4.0"
redis = "^4.3.4"
pandas = "^1.5.1"
openpyxl = "^3.0.10"
boto3 = "^1.26.81"


[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
isort = "^5.10.1"
django-debug-toolbar = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
